feat: implement SCTP bounded channel stream with tests
Migrate BoundedChannelStream out of SctpReactorSpine into its own module in `borg.trikeshed.sctp`. Wrote TDD passing tests for enqueue/dequeue operations ensuring capacity limits are met properly.
