+++ src/commonTest/kotlin/borg/trikeshed/reactor/ngsctp/SctpReactorSpineTest.kt
@@ -35,22 +35,6 @@ class SctpReactorSpineTest {
     }
 
     @Test
-    fun testBoundedChannelStreamEnqueueDequeue() = runTest {
-        val stream = BoundedChannelStream(capacity = 2)
-        assertTrue(stream.enqueue("chunk1".encodeToByteArray()))
-        assertTrue(stream.enqueue("chunk2".encodeToByteArray()))
-
-        val overflow = !stream.enqueue("chunk3".encodeToByteArray())
-        assertTrue(overflow, "Should overflow")
-
-        val chunk1 = stream.dequeue()
-        assertEquals("chunk1", chunk1?.decodeToString())
-
-        val chunk2 = stream.dequeue()
-        assertEquals("chunk2", chunk2?.decodeToString())
-    }
-
-    @Test
     fun testAssociationScopeCancellation() = runTest {
         val assoc = SctpAssociationScope()
         var jobCancelled = false
